requirements.txt
Flask==2.3.3
pyswisseph==2.10.3.2
timezonefinder==6.2.0
reportlab==4.0.4
Pillow==10.0.0
pytz==2023.3